contributing.md: Fix formatting
This commit contains two changes to CONTRIBUTING.md that:

Fix line line lengths. There were several instances where line lengths
were well over 80 characters. This commit adjusts them to make them
conform to formatting best practices, to stay under 80 characters when
possible.

Adhere to consist use of quotes. There are a few instances where smart
quotes are used (perhaps because it was pasted from a word processor).
This commit replaces them with dumb quotes to keep it consistent with
the quotation formatting found in the rest of the document.