Improve translation policy
This change separates the `make translate language=...` and `make
translate_base` Makefile targets. The former is useful for translators
and the latter for core developers.

This change also rearranges flags to `lupdate` to improve utility for
both cases.

Finally, this change also documents these targets and gives more advice
on how to use them.